+++ RELEASENOTES.md
@@ -1,5 +1,6 @@
 ## 0.2.0
 * RxJava - Original Future is not canceled upon unsubscribe.
+* guava-rxjava added
 
 ## 0.1.0
 * Artifact IDs renamed